+++ groups/etcd_group_service_test.go
@@ -20,48 +20,87 @@
 
 package groups
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/topfreegames/pitaya/v2/config"
+	"go.etcd.io/etcd/tests/v3/integration"
+)
+
+func setup(t *testing.T) (*integration.ClusterV3, GroupService) {
+	integration.BeforeTest(t)
+	cluster := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
+	cli := cluster.RandClient()
+	etcdGroupService, err := NewEtcdGroupService(*config.NewDefaultEtcdGroupServiceConfig(), cli)
+	if err != nil {
+		panic(err)
+	}
+
+	return cluster, etcdGroupService
+}
 
 func TestEtcdCreateDuplicatedGroup(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testCreateDuplicatedGroup(etcdGroupService, t)
 }
 
 func TestEtcdCreateGroup(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testCreateGroup(etcdGroupService, t)
 }
 
 func TestEtcdCreateGroupWithTTL(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testCreateGroupWithTTL(etcdGroupService, t)
 }
 
 func TestEtcdGroupAddMember(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testGroupAddMember(etcdGroupService, t)
 }
 
 func TestEtcdGroupAddDuplicatedMember(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testGroupAddDuplicatedMember(etcdGroupService, t)
 }
 
 func TestEtcdGroupContainsMember(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testGroupContainsMember(etcdGroupService, t)
 }
 
 func TestEtcdRemove(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testRemove(etcdGroupService, t)
 }
 
 func TestEtcdDelete(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testDelete(etcdGroupService, t)
 }
 
 func TestEtcdRemoveAll(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testRemoveAll(etcdGroupService, t)
 }
 
 func TestEtcdCount(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testCount(etcdGroupService, t)
 }
 
 func TestEtcdMembers(t *testing.T) {
+	cluster, etcdGroupService := setup(t)
+	defer cluster.Terminate(t)
 	testMembers(etcdGroupService, t)
 }
