+++ groups/memory_group_service_test.go
@@ -20,7 +20,20 @@
 
 package groups
 
-import "testing"
+import (
+	"os"
+	"testing"
+
+	"github.com/topfreegames/pitaya/v2/config"
+)
+
+var memoryGroupService *MemoryGroupService
+
+func TestMain(m *testing.M) {
+	memoryGroupService = NewMemoryGroupService(*config.NewDefaultMemoryGroupConfig())
+	exit := m.Run()
+	os.Exit(exit)
+}
 
 func TestMemoryCreateDuplicatedGroup(t *testing.T) {
 	testCreateDuplicatedGroup(memoryGroupService, t)
