+++ helpers/helpers.go
@@ -14,8 +14,8 @@ import (
 	"testing"
 	"time"
 
-	"go.etcd.io/etcd/clientv3"
-	"go.etcd.io/etcd/integration"
+	clientv3 "go.etcd.io/etcd/client/v3"
+	"go.etcd.io/etcd/tests/v3/integration"
 	"github.com/nats-io/nats-server/v2/server"
 	gnatsd "github.com/nats-io/nats-server/v2/test"
 )
@@ -64,6 +64,7 @@ func GetTestNatsServer(t *testing.T) *server.Server {
 // GetTestEtcd gets a test in memory etcd server
 func GetTestEtcd(t *testing.T) (*integration.ClusterV3, *clientv3.Client) {
 	t.Helper()
+	integration.BeforeTest(t)
 	c := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
 	cli := c.RandClient()
 	return c, cli
